+++ apps/restconf/restconf_lib.h
@@ -62,5 +62,6 @@ cbuf *readdata(FCGX_Request *r);
 int get_user_cookie(char *cookiestr, char  *attribute, char **val);
 int api_return_err(clicon_handle h, FCGX_Request *r, cxobj *xerr,
 		   int pretty, int use_xml);
+int restconf_terminate(clicon_handle h);
 
 #endif /* _RESTCONF_LIB_H_ */