+++ test/lib.sh
@@ -18,7 +18,7 @@ clixon_netconf=clixon_netconf
 
 # How to run restconf stand-alone and using valgrind
 #sudo su -c "/www-data/clixon_restconf -f $cfg -D 1" -s /bin/sh www-data
-#sudo su -c "valgrind --leak-check=full --show-leak-kinds=all /www-data/clixon_restconf -f $cfg -D 1" -s /bin/sh www-data
+#sudo su -c "valgrind --trace-children=no --child-silent-after-fork=yes --leak-check=full --show-leak-kinds=all /www-data/clixon_restconf -f $cfg -D 1" -s /bin/sh www-data
 
 #clixon_backend="valgrind --leak-check=full --show-leak-kinds=all clixon_backend"
 clixon_backend=clixon_backend